requirements.txt
Django==1.9.4
django-rest-swagger==0.3.4
djangorestframework==3.3.3
PyYAML==3.11
requests==2.9.1
